requirements.txt
requests
playwright
pillow